+++ src/api/sdk/common/mindsphere-sdk.ts
@@ -1,4 +1,5 @@
 import { AgentManagementClient } from "../agent/agent-management";
+import { AnomalyDetectionClient } from "../anomaly-detection/anomaly-detection";
 import { AssetManagementClient } from "../asset/asset-management";
 import { DataExchangeClient } from "../data-exchange/data-exchange";
 import { DataLakeClient } from "../data-lake/data-lake";
@@ -14,7 +15,6 @@ import { JobManagerClient } from "../jobmanager/jobmanager";
 import { KPICalculationClient } from "../kpi/kpi";
 import { MindConnectApiClient } from "../mcapi/mcapi";
 import { ModelManagementClient } from "../model/model-management";
-import { AnomalyDetectionClient } from "../anomaly-detection/anomaly-detection";
 import { NotificationClientV4 } from "../notification-v4/notification-v4";
 import { SemanticDataInterconnectClient } from "../sdi/sdi-v4";
 import { SignalCalculationClient } from "../signal-calculation/signal-calculation";
@@ -75,7 +75,7 @@ export class MindSphereSdk extends SdkClient {
      *
      * @memberOf MindSphereSdk
      */
-    public GetAnomalydetectionClient(): AnomalyDetectionClient {
+    public GetAnomalyDetectionClient(): AnomalyDetectionClient {
         this._anomalyDetectionClient = this._anomalyDetectionClient || new AnomalyDetectionClient(this._authenticator);
         return this._anomalyDetectionClient;
     }